+++ db-viewer-ui/stories/Themes/ObjectDetails.stories.tsx
@@ -1,4 +1,4 @@
-import { DEFAULT_THEME, DEFAULT_THEME_8PX, FLAT_THEME, ThemeContext, ThemeFactory } from "@skbkontur/react-ui";
+import { DEFAULT_THEME, DEFAULT_THEME_OLD, FLAT_THEME, ThemeContext, ThemeFactory } from "@skbkontur/react-ui";
 import { Theme } from "@skbkontur/react-ui/lib/theming/Theme";
 import React from "react";
 import StoryRouter from "storybook-react-router";
@@ -7,7 +7,6 @@ import { NullCustomRenderer } from "../../src";
 import { ObjectDetailsContainer } from "../../src/Containers/ObjectDetailsContainer";
 import { DbViewerApiFake } from "../Api/DbViewerApiFake";
 
-import { infraUiDark } from "./infraUiDark";
 import { reactUiDark } from "./reactUiDark";
 
 export default {
@@ -30,6 +29,5 @@ const DetailsContainer = ({ theme }: { theme: Theme }) => (
 
 export const Default = (): JSX.Element => <DetailsContainer theme={DEFAULT_THEME} />;
 export const Flat = (): JSX.Element => <DetailsContainer theme={FLAT_THEME} />;
-export const EightPx = (): JSX.Element => <DetailsContainer theme={DEFAULT_THEME_8PX} />;
+export const Old = (): JSX.Element => <DetailsContainer theme={DEFAULT_THEME_OLD} />;
 export const Dark = (): JSX.Element => <DetailsContainer theme={ThemeFactory.create(reactUiDark)} />;
-export const InfraDark = (): JSX.Element => <DetailsContainer theme={ThemeFactory.create(infraUiDark)} />;