+++ db-viewer-ui/stories/Themes/ObjectTypes.stories.tsx
@@ -1,4 +1,4 @@
-import { DEFAULT_THEME, DEFAULT_THEME_8PX, FLAT_THEME, ThemeContext, ThemeFactory } from "@skbkontur/react-ui";
+import { DEFAULT_THEME, DEFAULT_THEME_OLD, FLAT_THEME, ThemeContext, ThemeFactory } from "@skbkontur/react-ui";
 import { Theme } from "@skbkontur/react-ui/lib/theming/Theme";
 import React from "react";
 import StoryRouter from "storybook-react-router";
@@ -6,7 +6,6 @@ import StoryRouter from "storybook-react-router";
 import { ObjectTypesContainer } from "../../src/Containers/ObjectTypesContainer";
 import { DbViewerApiFake } from "../Api/DbViewerApiFake";
 
-import { infraUiDark } from "./infraUiDark";
 import { reactUiDark } from "./reactUiDark";
 
 export default {
@@ -27,6 +26,5 @@ const TypesContainer = ({ theme }: { theme: Theme }) => (
 
 export const Default = (): JSX.Element => <TypesContainer theme={DEFAULT_THEME} />;
 export const Flat = (): JSX.Element => <TypesContainer theme={FLAT_THEME} />;
-export const EightPx = (): JSX.Element => <TypesContainer theme={DEFAULT_THEME_8PX} />;
+export const Old = (): JSX.Element => <TypesContainer theme={DEFAULT_THEME_OLD} />;
 export const Dark = (): JSX.Element => <TypesContainer theme={ThemeFactory.create(reactUiDark)} />;
-export const InfraDark = (): JSX.Element => <TypesContainer theme={ThemeFactory.create(infraUiDark)} />;